Use a config object
